# More optimize for AArch32/AArch64