--- a/dlk/python/dlk/templates/cmake/toolchain/linux_arm.cmake
+++ b/dlk/python/dlk/templates/cmake/toolchain/linux_arm.cmake
@@ -4,7 +4,7 @@
 #
 set(TOOLCHAIN_PATH "")
 set(CMAKE_SYSROOT "")
-
+set(AARCH32 ON)
 
 #
 # Cross-compiling for ARM