--- a/dlk/python/dlk/templates/include/func/extract_image_patches.h
+++ b/dlk/python/dlk/templates/include/func/extract_image_patches.h
@@ -23,6 +23,10 @@ limitations under the License.
 #include "pack_input_to_qwords.h"
 #include <limits.h>
 
+#ifdef USE_NEON
+#include <arm_neon.h>
+#endif
+
 template <typename T>
 void func_ExtractImagePatches(
     const TensorView<T, MemoryLayout::NHWC>& input,
@@ -91,6 +95,23 @@ inline void func_ExtractImagePatches(
             T_UINT ch = (ki * kernel_size + kj) * bit_shift;
             T_UINT ch_high = ch / QUANTIZED_PACKED::BitCount;
             T_UINT ch_low = ch % QUANTIZED_PACKED::BitCount;
+#ifdef USE_NEON
+            const auto out_idx = wi * out_width * out_depth * bits_per_input
+              + wj * out_depth * bits_per_input
+              + ch_high * bits_per_input;
+            const auto in_idx = row * input_width * input_depth * bits_per_input
+              + col * input_depth * bits_per_input;
+            const auto in = vld1_u32(reinterpret_cast<uint32_t*>(input.data() + in_idx));
+            const auto masked = vand_u32(vdup_n_u32(mask), in);
+#ifdef AARCH32
+            const auto shifted = vshl_u32(masked, vdup_n_s32(ch_low));
+#else
+            const auto shifted = vshl_n_u32(masked, ch_low);
+#endif
+            const auto out_old = vld1_u32(reinterpret_cast<uint32_t*>(output.data() + out_idx));
+            const auto out_new = vorr_u32(out_old, shifted);
+            vst1_u32(reinterpret_cast<uint32_t*>(output.data() + out_idx), out_new);
+#else
             for(T_UINT digit = 0; digit < bits_per_input; ++digit) {
               const auto out_idx = wi * out_width * out_depth * bits_per_input
                 + wj * out_depth * bits_per_input
@@ -101,6 +122,7 @@ inline void func_ExtractImagePatches(
                 + digit;
               output.data()[out_idx] |= QUANTIZED_PACKED((mask & input.data()[in_idx].Raw()) << ch_low);
             }
+#endif
           }
   } else {
     for(T_UINT ih = 0; ih < input_depth; ++ih)
@@ -111,6 +133,17 @@ inline void func_ExtractImagePatches(
             {
               T_INT row = (wi * stride) + ki;
               T_INT col = (wj * stride) + kj;
+#ifdef USE_NEON
+              const auto ch_high = ih + (ki * kernel_size + kj) * input_depth;
+              const auto out_idx = wi * out_width * out_depth * bits_per_input
+                + wj * out_depth * bits_per_input
+                + ch_high * bits_per_input;
+              const auto in_idx = row * input_width * input_depth * bits_per_input
+                + col * input_depth * bits_per_input
+                + ih * bits_per_input;
+              const auto in = vld1_u32(reinterpret_cast<uint32_t*>(input.data() + in_idx));
+              vst1_u32(reinterpret_cast<uint32_t*>(output.data() + out_idx), in);
+#else
               for(T_UINT digit = 0; digit < bits_per_input; ++digit) {
                 const auto ch_high = ih + (ki * kernel_size + kj) * input_depth;
                 const auto out_idx = wi * out_width * out_depth * bits_per_input
@@ -124,6 +157,7 @@ inline void func_ExtractImagePatches(
                 output.data()[out_idx]
                   = input.data()[in_idx];
               }
+#endif
             }
   }
 
@@ -162,6 +196,23 @@ inline void func_ExtractImagePatches(
             T_UINT ch = (ki * kernel_size + kj) * bit_shift;
             T_UINT ch_high = ch / QUANTIZED_PACKED::BitCount;
             T_UINT ch_low = ch % QUANTIZED_PACKED::BitCount;
+#ifdef USE_NEON
+            const auto out_idx = ch_high * out_height * out_width * bits_per_input
+              + wi * out_width * bits_per_input
+              + wj * bits_per_input;
+            const auto in_idx = row * input_width * bits_per_input
+              + col * bits_per_input;
+            const auto in = vld1_u32(reinterpret_cast<uint32_t*>(input.data() + in_idx));
+            const auto masked = vand_u32(vdup_n_u32(mask), in);
+#ifdef AARCH32
+            const auto shifted = vshl_u32(masked, vdup_n_s32(ch_low));
+#else
+            const auto shifted = vshl_n_u32(masked, ch_low);
+#endif
+            const auto out_old = vld1_u32(reinterpret_cast<uint32_t*>(output.data() + out_idx));
+            const auto out_new = vorr_u32(out_old, shifted);
+            vst1_u32(reinterpret_cast<uint32_t*>(output.data() + out_idx), out_new);
+#else
             for(T_UINT digit = 0; digit < bits_per_input; ++digit) {
               const auto out_idx = ch_high * out_height * out_width * bits_per_input
                 + wi * out_width * bits_per_input
@@ -172,6 +223,7 @@ inline void func_ExtractImagePatches(
                 + digit;
               output.data()[out_idx] |= QUANTIZED_PACKED((mask & input.data()[in_idx].Raw()) << ch_low);
             }
+#endif
           }
   } else {
     for(T_UINT ih = 0; ih < input_depth; ++ih)
@@ -182,6 +234,17 @@ inline void func_ExtractImagePatches(
             {
               T_INT row = (wi * stride) + ki;
               T_INT col = (wj * stride) + kj;
+#ifdef USE_NEON
+              const auto ch_high = ih + (ki * kernel_size + kj) * input_depth;
+              const auto out_idx = ch_high * out_height * out_width * bits_per_input
+                + wi * out_width * bits_per_input
+                + wj * bits_per_input;
+              const auto in_idx = ih * input_height * input_width * bits_per_input
+                + row * input_width * bits_per_input
+                + col * bits_per_input;
+              const auto in = vld1_u32(reinterpret_cast<uint32_t*>(input.data() + in_idx));
+              vst1_u32(reinterpret_cast<uint32_t*>(output.data() + out_idx), in);
+#else
               for(T_UINT digit = 0; digit < bits_per_input; ++digit) {
                 const auto ch_high = ih + (ki * kernel_size + kj) * input_depth;
                 const auto out_idx = ch_high * out_height * out_width * bits_per_input
@@ -195,6 +258,7 @@ inline void func_ExtractImagePatches(
                 output.data()[out_idx]
                   = input.data()[in_idx];
               }
+#endif
             }
   }