<!--- 👉👉👉 Provide a general summary of your changes. 👈👈👈 -->
more faster on Arm CPUs!

| |DE10-Nano (CPU)|DE10-Nano (FPGA)|Raspberry Pi 3B|
|----|----|----|----|
|before|327ms|28.7ms|528ms|
|by #353 |129ms|28.7ms|83ms|
|after|~~97ms~~ 94ms|19.9ms|52ms|

## Motivation and Context
<!--- 💭💡💭 Why is this change required? What problem does it solve? 💭💡💭 -->
<!--- 🔗 If it fixes an open issue, please link to the issue here. 🔗 -->
After #353 , I found another optimization points.
This PR implement these.

## Description
<!--- 📝🎯📝 Describe your changes in detail. 📝🎯📝 -->
- Use pointer instead of TensorView::operator()
- Optimize extract_image_patches for Arm
- Fuse apply_thresholds and pack_16bit
- Allocate buffer in booting time

## How has this been tested?
<!--- 🙆👌🙆 Please describe in detail how you tested your changes. 🙆👌🙆 -->
<!--- Include details of your testing environment, details of your manual tests, snippet code or commands of your manual tests, table of experiments results metrics, tests ran to see how your change affects other areas of the code, etc. -->
<!--- If you have experiments report documents please link to here. -->

## Screenshots (if appropriate):

## Types of changes
<!--- 🏁🎌🏁 What types of changes does your code introduce? Put an `x` in all the boxes that apply: 🏁🎌🏁 -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature / Optimization (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)

## Checklist:
<!--- ✅✅✅ Go over all the following points, and put an `x` in all the boxes that apply. ✅✅✅ -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->

<!--- TODO(wakisaka): After decide our code style, add this.
- [ ] My code follows the code style of this project.
-->
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.