@primenumber 2019-07-10
Optimized for AArch32, 94ms on DE10-Nano!